tests: skip test part when missed in kernel

+++ test/shell/profiles-cache.sh
@@ -62,10 +62,12 @@ EOF
 aux prepare_vg 2 1000000
 
 # Check writecache read data from profile
+if aux have_writecache 1 0 0 ; then
 lvcreate -n $lv1 -l 4 -an $vg "$dev1"
 lvcreate -y --type writecache -l 4 --cachevol $lv1 -n $lv2 --metadataprofile $PFILE $vg "$dev2"
 check lv_field $vg/$lv2 cachesettings "high_watermark=60"
 lvremove -y $vg
+fi
 
 # Check chunk_size is grabbed from configuration
 lvcreate -L1G --config 'allocation/cache_pool_chunk_size=512' --type cache-pool $vg/cpool
